relocate KM model
